src/requirements/backend/requirements.txt
fastapi
pydantic
pydantic[email]
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
celery
flower
sqlalchemy-utils
pytz
tzlocal
